Refactored search methods to its own utility javascript. Made skp search debounce
